style(sudt): extract bootstrap from the main
